enrol and unenroll api latest changes (#43)
* enrol and unenroll api latest changes

+++ course-mw/enrolment-actor/src/main/scala/org/sunbird/enrolments/CourseEnrolmentActor.scala
@@ -82,12 +82,17 @@ class CourseEnrolmentActor @Inject()(@Named("course-batch-notification-actor") c
         val enrolmentData: UserCourses = userCoursesDao.read(request.getRequestContext, userId, courseId, batchId)
         validateEnrolment(batchData, enrolmentData, true)
         val data: java.util.Map[String, AnyRef] = createUserEnrolmentMap(userId, courseId, batchId, enrolmentData, request.getContext.getOrDefault(JsonKey.REQUEST_ID, "").asInstanceOf[String])
-        upsertEnrollment(userId, courseId, batchId, data, (null == enrolmentData), request.getRequestContext)
-        logger.info(request.getRequestContext, "CourseEnrolmentActor :: enroll :: Deleting redis for key " + getCacheKey(userId))
-        cacheUtil.delete(getCacheKey(userId))
-        sender().tell(successResponse(), self)
-        generateTelemetryAudit(userId, courseId, batchId, data, "enrol", JsonKey.CREATE, request.getContext)
-        notifyUser(userId, batchData, JsonKey.ADD)
+        val hasAccess = validateCourseSecureSettings(courseId, request)
+        if (hasAccess) {
+            upsertEnrollment(userId, courseId, batchId, data, (null == enrolmentData), request.getRequestContext)
+            logger.info(request.getRequestContext, "CourseEnrolmentActor :: enroll :: Deleting redis for key " + getCacheKey(userId))
+            cacheUtil.delete(getCacheKey(userId))
+            sender().tell(successResponse(), self)
+            generateTelemetryAudit(userId, courseId, batchId, data, "enrol", JsonKey.CREATE, request.getContext)
+            notifyUser(userId, batchData, JsonKey.ADD)
+        } else {
+            ProjectCommonException.throwClientErrorException(ResponseCode.accessDeniedToEnrolOrUnenrolCourse, courseId)
+        }
     }
     
     
@@ -100,12 +105,17 @@ class CourseEnrolmentActor @Inject()(@Named("course-batch-notification-actor") c
         getUpdatedStatus(enrolmentData)
         validateEnrolment(batchData, enrolmentData, false)
         val data: java.util.Map[String, AnyRef] = new java.util.HashMap[String, AnyRef]() {{ put(JsonKey.ACTIVE, ProjectUtil.ActiveStatus.INACTIVE.getValue.asInstanceOf[AnyRef]) }}
-        upsertEnrollment(userId,courseId, batchId, data, false, request.getRequestContext)
-        logger.info(request.getRequestContext, "CourseEnrolmentActor :: unEnroll :: Deleting redis for key " + getCacheKey(userId))
-        cacheUtil.delete(getCacheKey(userId))
-        sender().tell(successResponse(), self)
-        generateTelemetryAudit(userId, courseId, batchId, data, "unenrol", JsonKey.UPDATE, request.getContext)
-        notifyUser(userId, batchData, JsonKey.REMOVE)
+        val hasAccess = validateCourseSecureSettings(courseId, request)
+        if (hasAccess) {
+            upsertEnrollment(userId, courseId, batchId, data, false, request.getRequestContext)
+            logger.info(request.getRequestContext, "CourseEnrolmentActor :: unEnroll :: Deleting redis for key " + getCacheKey(userId))
+            cacheUtil.delete(getCacheKey(userId))
+            sender().tell(successResponse(), self)
+            generateTelemetryAudit(userId, courseId, batchId, data, "unenrol", JsonKey.UPDATE, request.getContext)
+            notifyUser(userId, batchData, JsonKey.REMOVE)
+        } else {
+            ProjectCommonException.throwClientErrorException(ResponseCode.accessDeniedToEnrolOrUnenrolCourse, courseId)
+        }
     }
 
     def list(request: Request): Unit = {
@@ -344,6 +354,49 @@ class CourseEnrolmentActor @Inject()(@Named("course-batch-notification-actor") c
             contents.get(0).asInstanceOf[java.util.Map[String, AnyRef]].getOrDefault(JsonKey.LEAF_NODE_COUNT, 0.asInstanceOf[AnyRef]).asInstanceOf[Int]
         } else 0}
         enrolmentData.setStatus(getCompletionStatus(enrolmentData.getProgress, leafNodesCount))
+    }
+
+    private def validateCourseSecureSettings(courseId: String, request: Request): Boolean = {
+        val requestBody: String = prepareContentSearchRequest(courseId, request)
+        val searchResult: java.util.Map[String, AnyRef] = ContentSearchUtil.searchContentSync(request.getRequestContext, request.getContext.getOrDefault(JsonKey.URL_QUERY_STRING, "").asInstanceOf[String], requestBody, request.get(JsonKey.HEADER).asInstanceOf[java.util.Map[String, String]])
+        val coursesList: java.util.List[java.util.Map[String, AnyRef]] = searchResult.getOrDefault(JsonKey.CONTENTS, new java.util.ArrayList[java.util.Map[String, AnyRef]]()).asInstanceOf[java.util.List[java.util.Map[String, AnyRef]]]
+        val organizationIds = new java.util.ArrayList[String]()
+        coursesList.asScala.foreach { course =>
+            val secureSettingsOption = Option(course.get("secureSettings").asInstanceOf[java.util.Map[String, AnyRef]])
+            secureSettingsOption.foreach { secureSettings =>
+                Option(secureSettings.get("organisation")) match {
+                    case Some(ids: java.util.List[String]) =>
+                        organizationIds.addAll(ids)
+                    case _ =>
+                }
+            }
+        }
+        val userChannelId: String = request.getRequest.getOrDefault("x-user-channel-id", "").asInstanceOf[String]
+        val hasAccess = organizationIds.isEmpty || organizationIds.contains(userChannelId)
+        hasAccess
+    }
+
+    private def prepareContentSearchRequest(courseIds: String, request: Request): String = {
+        val filters: java.util.Map[String, AnyRef] = new java.util.HashMap[String, AnyRef]() {
+            {
+                put(JsonKey.IDENTIFIER, java.util.Collections.singletonList(courseIds))
+                put(JsonKey.STATUS, "Live")
+                put(JsonKey.MIME_TYPE, JsonKey.COLLECTION_MIME_TYPE)
+                put(JsonKey.TRACKABLE_ENABLED, JsonKey.YES)
+                putAll(request.getRequest.getOrDefault(JsonKey.FILTERS, new java.util.HashMap[String, AnyRef]).asInstanceOf[java.util.Map[String, AnyRef]])
+            }
+        }
+        val searchRequest: java.util.Map[String, java.util.Map[String, AnyRef]] = new java.util.HashMap[String, java.util.Map[String, AnyRef]]() {
+            {
+                put(JsonKey.REQUEST, new java.util.HashMap[String, AnyRef]() {
+                    {
+                        put(JsonKey.FILTERS, filters)
+                        put(JsonKey.LIMIT, 1.asInstanceOf[AnyRef])
+                    }
+                })
+            }
+        }
+        new ObjectMapper().writeValueAsString(searchRequest)
     }
 }
